Reorganizing game flow
