--- Assets/Scenes/Level1.unity.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: e8d5ad5474ca24adda5a118f80450cba
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
