+++ Assets/Scenes/LevelTwoScene.unity
@@ -513,6 +513,63 @@ Transform:
   m_CorrespondingSourceObject: {fileID: 5408656516596256000, guid: 7986fe9d316fc83429e12b8d33f355ff, type: 3}
   m_PrefabInstance: {fileID: 232426972}
   m_PrefabAsset: {fileID: 0}
+--- !u!1001 &282158847
+PrefabInstance:
+  m_ObjectHideFlags: 0
+  serializedVersion: 2
+  m_Modification:
+    serializedVersion: 3
+    m_TransformParent: {fileID: 0}
+    m_Modifications:
+    - target: {fileID: 1760078745661375574, guid: f4126d06fbb6f4899b70c8d9b489b59b, type: 3}
+      propertyPath: m_LocalPosition.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 1760078745661375574, guid: f4126d06fbb6f4899b70c8d9b489b59b, type: 3}
+      propertyPath: m_LocalPosition.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 1760078745661375574, guid: f4126d06fbb6f4899b70c8d9b489b59b, type: 3}
+      propertyPath: m_LocalPosition.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 1760078745661375574, guid: f4126d06fbb6f4899b70c8d9b489b59b, type: 3}
+      propertyPath: m_LocalRotation.w
+      value: 1
+      objectReference: {fileID: 0}
+    - target: {fileID: 1760078745661375574, guid: f4126d06fbb6f4899b70c8d9b489b59b, type: 3}
+      propertyPath: m_LocalRotation.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 1760078745661375574, guid: f4126d06fbb6f4899b70c8d9b489b59b, type: 3}
+      propertyPath: m_LocalRotation.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 1760078745661375574, guid: f4126d06fbb6f4899b70c8d9b489b59b, type: 3}
+      propertyPath: m_LocalRotation.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 1760078745661375574, guid: f4126d06fbb6f4899b70c8d9b489b59b, type: 3}
+      propertyPath: m_LocalEulerAnglesHint.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 1760078745661375574, guid: f4126d06fbb6f4899b70c8d9b489b59b, type: 3}
+      propertyPath: m_LocalEulerAnglesHint.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 1760078745661375574, guid: f4126d06fbb6f4899b70c8d9b489b59b, type: 3}
+      propertyPath: m_LocalEulerAnglesHint.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 4333145006690809556, guid: f4126d06fbb6f4899b70c8d9b489b59b, type: 3}
+      propertyPath: m_Name
+      value: AudioManager
+      objectReference: {fileID: 0}
+    m_RemovedComponents: []
+    m_RemovedGameObjects: []
+    m_AddedGameObjects: []
+    m_AddedComponents: []
+  m_SourcePrefab: {fileID: 100100000, guid: f4126d06fbb6f4899b70c8d9b489b59b, type: 3}
 --- !u!1 &287899293
 GameObject:
   m_ObjectHideFlags: 0
@@ -2140,6 +2197,53 @@ Tilemap:
     e31: 0
     e32: 0
     e33: 1
+--- !u!1 &371590017
+GameObject:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  serializedVersion: 6
+  m_Component:
+  - component: {fileID: 371590019}
+  - component: {fileID: 371590018}
+  m_Layer: 0
+  m_Name: LevelLoader
+  m_TagString: Untagged
+  m_Icon: {fileID: 0}
+  m_NavMeshLayer: 0
+  m_StaticEditorFlags: 0
+  m_IsActive: 1
+--- !u!114 &371590018
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 371590017}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: a291d3850acc4a3488f590ab18d654ad, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  levelName: 
+  playerSpawnPosition: {x: 0, y: 0}
+  respawnPlayerAtPosition: 1
+--- !u!4 &371590019
+Transform:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 371590017}
+  serializedVersion: 2
+  m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
+  m_LocalPosition: {x: 0, y: 0, z: 0}
+  m_LocalScale: {x: 1, y: 1, z: 1}
+  m_ConstrainProportionsScale: 0
+  m_Children: []
+  m_Father: {fileID: 0}
+  m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
 --- !u!1 &432688228
 GameObject:
   m_ObjectHideFlags: 0
@@ -14092,3 +14196,5 @@ SceneRoots:
   - {fileID: 1957793239}
   - {fileID: 898279742}
   - {fileID: 1496010376}
+  - {fileID: 282158847}
+  - {fileID: 371590019}
