+++ Assets/Scenes/MainLevelScene.unity
@@ -16050,6 +16050,53 @@ Transform:
   m_CorrespondingSourceObject: {fileID: 8041078681807074851, guid: dbd0cb661d4232c47bfb265ef9f006e8, type: 3}
   m_PrefabInstance: {fileID: 862219824}
   m_PrefabAsset: {fileID: 0}
+--- !u!1 &886347783
+GameObject:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  serializedVersion: 6
+  m_Component:
+  - component: {fileID: 886347784}
+  - component: {fileID: 886347785}
+  m_Layer: 0
+  m_Name: LevelLoader
+  m_TagString: Untagged
+  m_Icon: {fileID: 0}
+  m_NavMeshLayer: 0
+  m_StaticEditorFlags: 0
+  m_IsActive: 1
+--- !u!4 &886347784
+Transform:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 886347783}
+  serializedVersion: 2
+  m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
+  m_LocalPosition: {x: 0, y: 0, z: 0}
+  m_LocalScale: {x: 1, y: 1, z: 1}
+  m_ConstrainProportionsScale: 0
+  m_Children: []
+  m_Father: {fileID: 0}
+  m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
+--- !u!114 &886347785
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 886347783}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: a291d3850acc4a3488f590ab18d654ad, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  levelName: 
+  playerSpawnPosition: {x: 0, y: 0}
+  respawnPlayerAtPosition: 1
 --- !u!1 &901920880
 GameObject:
   m_ObjectHideFlags: 0
@@ -18361,3 +18408,4 @@ SceneRoots:
   - {fileID: 2079813600}
   - {fileID: 9144459157120408083}
   - {fileID: 2132084385}
+  - {fileID: 886347784}
