--- Assets/Scenes/test Alex 1.unity.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: bc5f594dd9d969c4b9fd7c4e8f9251a0
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
