--- Assets/Scripts/MainMenu.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 222acd3e2bdcf47f6ba518d35c7933d2
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
