--- Assets/Scripts/MainMenu/MainMenu.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 6196ab30c31b44f37b64aa24a8865682
